Mix up int types in convert test

+++ test/utils/ConvertTest.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <gtest/gtest.h>
 
 #include "../../src/utils/Convert.h"
@@ -14,7 +15,7 @@ TEST_F(ConverterTest, WillConvertValidInteger)
     std::string input = "42";
 
     // WHEN
-    int result = Converter::convert<int>(input);
+    int result = Converter::convert<uint16_t>(input);
 
     // THEN
     EXPECT_EQ(result, 42);
@@ -50,7 +51,7 @@ TEST_F(ConverterTest, WillConvertZero)
     std::string input = "0";
 
     // WHEN
-    int result = Converter::convert<int>(input);
+    int result = Converter::convert<int32_t>(input);
 
     // THEN
     EXPECT_EQ(result, 0);
@@ -62,7 +63,7 @@ TEST_F(ConverterTest, WillConvertWithWhitespace)
     std::string input = "   99  ";
 
     // WHEN
-    int result = Converter::convert<int>(input);
+    int result = Converter::convert<uint64_t>(input);
 
     // THEN
     EXPECT_EQ(result, 99);
